requirements.txt
# requirements.txt
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0

# Data processing
pandas>=2.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
streamlit>=1.28.0  # Streamlit pour l'interface web

# Logging & Monitoring
tensorboard>=2.13.0
tqdm>=4.65.0
wandb>=0.15.0  # optionnel

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
albumentations>=1.3.0  # augmentations avancées
timm>=0.9.0  # modèles pré-entraînés

# Additional utilities for multimodal AI
transformers>=4.30.0  # pour le traitement du texte
sentence-transformers>=2.2.0  # embeddings de texte